Fix to results being reset after pagination update.
1. Call to getInterpreterContext().out().clear() in DTTableDatasetNg causes a reset to Paragraph.results.

2. getInterpreterContext().out().flush() on the other hand does not update Paragraph.results, it updates Paragraph.outputBuffer.

3. The data sent to ui comes from Paragraph.results, which was previously cleared, leading to results being dropped.

When running a paragraph, the Paragraph.result is set by Job.completeWithSuccess() -> Job.setResult() method, which is not used when pagination is changed.

Since the source of truth in Zeppelin is the state of the Note object in memory, empty Paragraph.result also is saved to disk when another change is made.

This fix works by providing a boolean to the already defined getInterpreterContext().out().clear() method.
Providing true clears the results (wanted behaviour when paragraph is executed)
Providing false skips the call to clear results (wanted behaviour when changing pagination)